allow for HLS playlists
